requirements.txt
Pillow
cozmo
anki-vector==0.5.1
